requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.10
django-crispy-forms==1.12.0
django-quill-editor==0.1.22
django-utils-six==2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==19.9.0
openpyxl==3.0.9
Pillow==8.0.1
psycopg2==2.9.1
python-decouple==3.4
python-http-client==3.3.4
pytz==2021.3
raven==6.10.0
sendgrid==6.8.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
whitenoise==4.1.2
